Refactored trees file into a new subdir.
